gestion user

+++ controllers/c_deleteImage.php
@@ -1,11 +1,17 @@
 <?php
-require_once(PATH_MODELS.'ImagesDAO.php');
-$imagesDAO = new ImagesDAO();
 
-if(isset($_GET['id'])){
-  $id = htmlspecialchars($_GET['id']);
-  $imagesDAO->deleteImageById($id);
+if(isset($_SESSION['connect'])){
 
-  header('location: ../?page=edit');
+  require_once(PATH_MODELS.'ImagesDAO.php');
+  $imagesDAO = new ImagesDAO();
 
+  if(isset($_GET['id'])){
+    $id = htmlspecialchars($_GET['id']);
+    $imagesDAO->deleteImageById($id);
+
+    header('location: ../?page=edit');
+
+  }
+} else {
+  header('location: ../?page=connexion');
 }
