+++ controllers/c_deleteModele.php
@@ -1,11 +1,17 @@
 <?php
-require_once(PATH_MODELS.'ModelesDAO.php');
-$modelesDAO = new ModelesDAO();
 
-if(isset($_GET['id'])){
-  $id = htmlspecialchars($_GET['id']);
-  $modelesDAO->deleteModeleById($id);
+if(isset($_SESSION['connect'])){
 
-  header('location: ../?page=editModele');
+  require_once(PATH_MODELS.'ModelesDAO.php');
+  $modelesDAO = new ModelesDAO();
 
+  if(isset($_GET['id'])){
+    $id = htmlspecialchars($_GET['id']);
+    $modelesDAO->deleteModeleById($id);
+
+    header('location: ../?page=editModele');
+
+  }
+} else {
+  header('location: ../?page=connexion');
 }
